+++ History.md
@@ -2,6 +2,7 @@
 ==================
 
 * Bumps clean-css dependency to 5.0.
+* Bumps commander dependency to 7.0.
 
 [4.3.0 / 2019-04-06](https://github.com/jakubpawlowicz/clean-css-cli/compare/4.2...v4.3.0)
 ==================